refactor: Organize codebase and optimization
