+++ Cargo.toml
@@ -43,18 +43,23 @@ tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 tempfile = "3.17.1"
 
 [profile.release]
-opt-level = 3
 lto = "fat"
 codegen-units = 1
-panic = 'abort'
+panic = "abort"
 strip = true
+opt-level = 3
 debug = false
 incremental = false
 overflow-checks = false
+rpath = false
+debug-assertions = false
 
 [profile.dev]
-opt-level = 0
-debug = true
+debug-assertions = true
+overflow-checks = true
+
+[package.metadata]
+codegen-units-std = 1
 
 # The profile that 'cargo dist' will build with
 [profile.dist]
